requirements.txt
Flask==2.3.3
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.0.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
# Enhanced OCR dependencies
opencv-python==4.8.1.78
numpy==1.24.3
# Optional for PDF to image conversion
PyMuPDF==1.23.8
hashlib-compat==1.0.1
python-magic-bin==0.4.14
requests==2.31.0